Project-01.DBUHub/FastAPI-Admin-Backend/requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.7.2
asyncpg==0.29.0
attrs==23.2.0
bcrypt==3.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
coverage==7.4.4
crayons==0.4.0
cryptography==42.0.8
demjson3==3.0.6
distlib==0.3.8
Django==5.0.2
django-environ==0.11.2
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
environ==1.0
facebook_page_scraper==5.0.6
fake-useragent==1.5.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
kaitaistruct==0.10
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.6
outcome==1.3.0.post0
parse==1.20.2
passlib==1.7.4
pkginfo==1.11.1
platformdirs==4.2.2
poetry-core==1.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydivert==2.1.0
pygame==2.5.2
Pygments==2.18.0
pymongo==4.8.0
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypika-tortoise==0.1.6
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
rich==13.7.1
rsa==4.9
selenium==4.1.0
selenium-wire==5.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
sqlparse==0.4.3
starlette==0.37.2
tomlkit==0.13.0
tortoise-orm==0.21.5
tqdm==4.66.4
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2024.7.2
typer==0.12.3
typing_extensions==4.9.0
tzdata==2022.6
tzlocal==5.2
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.12
urllib3-secure-extra==0.1.0
uvicorn==0.27.1
w3lib==2.2.1
watchfiles==0.22.0
webdriver-manager==3.2.2
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
